move Protocol negosiation handler to http2 package
